requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: osx-64
blas=1.0=mkl
blinker=1.4=py37hecd8cb5_0
brotlipy=0.7.0=py37h9ed2024_1003
ca-certificates=2021.1.19=hecd8cb5_0
cachetools=4.2.1=pyhd3eb1b0_0
certifi=2020.12.5=py37hecd8cb5_0
cffi=1.14.4=py37h2125817_0
chardet=4.0.0=py37hecd8cb5_1003
click=7.1.2=pyhd3eb1b0_0
cryptography=3.3.1=py37hbcfaee0_0
google-api-core=1.22.2=py37_0
google-api-python-client=1.12.8=pyhd3deb0d_0
google-auth=1.21.3=py_0
google-auth-httplib2=0.0.4=pyh9f0ad1d_0
google-auth-oauthlib=0.4.2=pyhd3eb1b0_2
googleapis-common-protos=1.52.0=py37_0
httplib2=0.18.1=pyh9f0ad1d_0
idna=2.10=pyhd3eb1b0_0
importlib-metadata=2.0.0=py_1
importlib_metadata=2.0.0=1
intel-openmp=2019.4=233
keyring=22.0.1=py37hecd8cb5_0
krb5=1.17.1=hddcf347_0
libcxx=10.0.0=1
libedit=3.1.20191231=h1de35cc_1
libffi=3.3=hb1e8313_2
libpq=12.2=h051b688_0
libprotobuf=3.13.0.1=hab81aa3_0
mkl=2019.4=233
mkl-service=2.3.0=py37h9ed2024_0
mkl_fft=1.2.0=py37hc64f4ea_0
mkl_random=1.1.1=py37h959d312_0
ncurses=6.2=h0a44026_1
numpy=1.19.2=py37h456fd55_0
numpy-base=1.19.2=py37hcfb5961_0
oauthlib=3.1.0=py_0
openssl=1.1.1i=h9ed2024_0
pip=20.3.3=py37hecd8cb5_0
protobuf=3.13.0.1=py37hb1e8313_1
psycopg2=2.8.5=py37hbcfaee0_1
pyasn1=0.4.8=py_0
pyasn1-modules=0.2.8=py_0
pycparser=2.20=py_2
pyjwt=1.7.1=py37_0
pyopenssl=20.0.1=pyhd3eb1b0_1
pyscaffold=3.3.1=pyhd8ed1ab_0
pysocks=1.7.1=py37hecd8cb5_0
python=3.7.9=h26836e1_0
python_abi=3.7=1_cp37m
pytz=2021.1=pyhd3eb1b0_0
readline=8.0=h1de35cc_0
requests=2.25.1=pyhd3eb1b0_0
requests-oauthlib=1.3.0=py_0
rsa=4.7=pyhd3eb1b0_1
setuptools=52.0.0=py37hecd8cb5_0
simplejson=3.17.2=py37haf1e3a3_0
six=1.15.0=py37hecd8cb5_0
sqlite=3.33.0=hffcf06c_0
tableauserverclient=0.14.1=pyhd8ed1ab_0
tk=8.6.10=hb0a8c7a_0
uritemplate=3.0.1=py_0
urllib3=1.26.3=pyhd3eb1b0_0
wheel=0.36.2=pyhd3eb1b0_0
xz=5.2.5=h1de35cc_0
zipp=3.4.0=pyhd3eb1b0_0
zlib=1.2.11=h1de35cc_3
